Turn dusk-merkle in a workspace
- merkle tree implementation with poseidon hash is under
  'poseidon-merkle'
- merkle tree implementation with blake3 included as an example
- in poseidon-merkle plonk is updated to 0.14
- in poseidon-merkle poseidon is updated to 0.30
