+++ dusk-merkle/src/node.rs
