+++ dusk-merkle/src/opening.rs
