+++ dusk-merkle/tests/serializable_opening.rs
