+++ poseidon-merkle/CHANGELOG.md
@@ -1,0 +1,21 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+[0.1.0] - 2023-06-28
+
+## Added
+
+- Add poseidon-merkle crate [#58]
+
+<!-- ISSUES -->
+[#58]: https://github.com/dusk-network/merkle/issues/58
+
+<!-- VERSIONS -->
+[Unreleased]: https://github.com/dusk-network/merkle/compare/poseidon-merkle_v0.1.0...HEAD
+[0.1.0]: https://github.com/dusk-network/merkle/releases/tag/poseidon-merkle_v0.1.0
