+++ poseidon-merkle/Cargo.toml
@@ -1,0 +1,57 @@
+[package]
+name = "poseidon-merkle"
+description = "Crate implementing Dusk Network's Merkle tree"
+version = "0.1.0"
+
+categories = ["data-structures", "no-std"]
+keywords = ["tree", "merkle", "poseidon", "hash", "data", "structure"]
+repository = "https://github.com/dusk-network/merkle"
+authors = [
+    "Eduardo Leegwater Simões <eduardo@dusk.network>",
+    "Moana Marcello <moana@dusk.network>",
+]
+
+edition = "2021"
+license = "MPL-2.0"
+
+[dependencies]
+dusk-merkle = "0.4"
+dusk-poseidon = { version = "0.30", default-features = false, features = ["merkle"] }
+dusk-bls12_381 = { version = "0.11", default-features = false }
+dusk-plonk = { version = "0.14", optional = true, default-features = false }
+rkyv = { version = "0.7", optional = true, default-features = false }
+bytecheck = { version = "0.6", optional = true, default-features = false }
+
+[dev-dependencies]
+rand = "0.8"
+criterion = "0.3"
+
+[features]
+zk = ["dusk-plonk/alloc", "dusk-poseidon/alloc"]
+size_16 = ["rkyv/size_16"]
+size_32 = ["rkyv/size_32"]
+size_64 = ["rkyv/size_64"]
+rkyv-impl = [
+    "rkyv/validation",
+    "rkyv/alloc",
+    "rkyv",
+    "bytecheck",
+    "dusk-bls12_381/rkyv-impl"
+]
+
+[[bench]]
+name = "poseidon"
+harness = false
+
+[[bench]]
+name = "zk"
+harness = false
+required-features = ["zk"]
+
+[[test]]
+name = "zk"
+required-features = ["zk"]
+
+[[example]]
+name = "zk"
+required-features = ["zk"]
